Project/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multitasking==0.0.11
mysql-connector-python==9.3.0
mysql-connector-repackaged==0.3.1
numpy==2.2.5
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yfinance==0.2.56
